Transfer contents of ups products ubtools, ubfcl, and ubxml into uboonecode.
Fcl files are in the source tree under uboonecode/fcl and are installed in
uboonecode/job.

Xml files are in the source tree under uboonecode/xml and are installed in
uboonecode/xml.

Executable scripts are in the source tree under uboonecode/scripts and are
installed in uboonecode_fq/bin.

Python modules are in the source tree under uboonecode/python and are
installed in uboonecode/python.